Rename property grid control factory and item factory to PropertyGridControlFactory and PropertyGridOperator. More consistent with the DataGrid...
