+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 - DataGrid: IDataGridControlFactory interface
 - DataGrid: TemplateColumnDefinition binding path
 - Target .NET 4.5 #117
+- PropertyGrid: Renamed PropertyControlFactory to PropertyGridControlFactory
+- PropertyGrid: Renamed PropertyItemFactory to PropertyGridOperator
 
 ### Added
 - DataGrid: Support for different types in the same column #118
