--- Source/Examples/PropertyGrid/PropertyGridDemo/LocalPropertyItemFactory.cs
@@ -1,14 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="LocalPropertyItemFactory.cs" company="PropertyTools">
-//   Copyright (c) 2014 PropertyTools contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace PropertyGridDemo
-{
-    using PropertyTools.Wpf;
-
-    public class LocalPropertyItemFactory : DefaultPropertyItemFactory
-    {
-    }
-}
